Do not modify the input markdown file
It's a little bit surprising for a conversion tool to modify the input data.
Rather than writing the results of each regex substitution back to disk, as
`replace` did, we store the results in intermediate variables, and only write
to disk when we need to.

I expect this leads to a small performance improvement.
It also means we can split out the text editing parts and write tests.